added dry run

+++ jme/stagecache/text_metadata.py
@@ -56,16 +56,20 @@ class Lockable():
         self.umask = cache.config['cache_umask']
         self.umask_dir = self.umask + 0o111
 
-    def get_write_lock(self, sleep_interval=3, force=False):
+    def get_write_lock(self, sleep_interval=3, force=False, dry_run=False):
         """ mark file as in progress (wait for existing lock) """
         if os.path.exists(self.write_lock):
             if force:
                 os.remove(self.write_lock)
+            if dry_run:
+                return
             logging.info('Waiting for lock...')
             logging.debug("force is "+ str(force))
             while os.path.exists(self.write_lock):
                 time.sleep(sleep_interval)
 
+        if dry_run:
+            return
         with open(self.write_lock, 'wt') as LOCK:
             LOCK.write('locked')
         os.chmod(self.write_lock, self.umask)
